+++ CHANGELOG.md
@@ -2945,3 +2945,4 @@
 - [2026-07-20 11:22] feat(i18n): implement responsive i18n view component
 - [2026-07-20 12:05] perf(perf): optimize memory consumption in perf list viewport
 - [2026-07-20 12:22] feat(cache): add bi-directional event bus handler for cache
+- [2026-07-20 13:16] feat(core): implement shimmer loading skeleton in core